requirements.txt
#django==1.10.1
#django-widget-tweaks
django_static_jquery
paho-mqtt
psycopg2
django-tastypie
pillow
simplejson
defusedxml
lxml
django-multiselectfield
